Remote API error handling: expose response string
Extend error handling by retrieving and providing the
response string from the remote API as part of the error
message returned to the caller if the status code
indicates an error condition.

Based on testing, the response string contains useful
troubleshooting information which can help identify
missing information in the submitted MessageCard payload.

refs GH-41