Replace IsType assert with Go 1.13 error wrapping
Update table test to filter error result (nil or not) and
then apply errors.As() func where non-nil to handle wrapped
errors.

I'm not particularly practiced at writing tests, so there is
likely a better way to handle this.

This solves the immediate problem, but more work is needed to
account for recent changes that I introduced with prior PRs
accepted upstream (and now locally).

refs GH-23