TestTeamsClientSend | Explicitly set response body
Explicitly set the http.Response.Body field. This prevents
a nil pointer dereference in versions of Go prior to 1.15.

Credit: Further review of the guide provided by @hassansine
that lead to the original tests written by @dasrick.

@dasrick had a different need for the implementation of the
original tests for this project, so implementing the Body field
at that time was not necessary.

With the changes in GH-43 to *use* the Body field for
diagnostic and client code feedback purposes, the tests needed
to be extended to handle setting this field to a relevant
value.

---

refs GH-43
refs http://hassansin.github.io/Unit-Testing-http-client-in-Go
refs 2ce144fffe3eab0da3282a80bc2a5ebdd58f1a42