Extend validation to assert specific response text
Per @stmoody's report in the `rveachkc/pymsteams` project,
Microsoft Teams webhook endpoint(s) will return a 200 status
response code in some scenarios when a message was actually
(seemingly silently) rejected due to size limits. Microsoft
Teams developers shared that a response string of `1` must
be present alongside a 200 status response code to indicate
successful message submission by the client.

This commit extends response validation to check for the
presence of this specific response string, otherwise returns
a custom error along with a description of what was expected
and what was actually received.

refs rveachkc/pymsteams#75
refs GH-59