Extend MessageCardPotentialAction validation
Add two helper functions to handle:

- validation of `*MessageCardPotentialAction` values
- appending `*MessageCardPotentialAction` values to the
  `PotentialAction` field for both `MessageCard` and
  `MessageCardSection` values

The `NewMessageCardPotentialAction` factory function has been
modified to return an `error`. This provides validation at both
entry points:

- creating a new value
- adding a value to an existing collection

**NOTE:**

This change is considered to be compatible with semantic
versioning because the support for `PotentialActions` has not
yet landed in a stable public release, leaving the final form
subject to change until the next public release.

refs GH-102